% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B_BIC.R
\name{B_BIC}
\alias{B_BIC}
\title{Bayesian Information Criteria for parameter selection.}
\usage{
B_BIC(Y, X, B, tuning, linear = FALSE)
}
\arguments{
\item{Y}{The index matrix}

\item{X}{The covariate matrix}

\item{B}{The initial coefficient matrix}

\item{tuning}{A coefficient matrix that need to be tuned.}

\item{linear}{Whether the link function is linear or not. The default value is False.}
}
\value{
A list that contains information about the Log Mean Square Error and the Degree of Freedom.
}
\description{
Bayesian Information Criteria for parameter selection.
}
