% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generate.R
\name{data_generate}
\alias{data_generate}
\title{Generate simulate data.}
\usage{
data_generate(fun.list, n, p, B.true, snr, cor = 0)
}
\arguments{
\item{fun.list}{A list of linear and nonlinear function for link function.}

\item{n}{The number of sample in the generated data.}

\item{p}{The number of predictors in the generated data.}

\item{B.true}{The groundtruth of the coefficient matrix B.}

\item{snr}{The Signal Noise Ratio of data generation.}

\item{cor}{The corvariance of the generate data. Default value is 0.}
}
\description{
Generate simulate data.
}
