% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env_check.R
\name{env_check}
\alias{env_check}
\title{check the environment for necessary packages, and install the missing packages}
\usage{
env_check()
}
\description{
check the environment for necessary packages, and install the missing packages
}
\examples{
env_check()
}
