% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_B_ADMM.R
\name{get_B_ADMM}
\alias{get_B_ADMM}
\title{The main function of mSIM package}
\usage{
get_B_ADMM(
  Y,
  X,
  B = NULL,
  lambda,
  rank,
  alpha = 1,
  control1 = list(max.iter = 100, tol = 0.01),
  control2 = list(ele.sparse = FALSE, row.sparse = TRUE, low.rank = TRUE),
  select.method = "linear",
  descent.method = "bfgs",
  plot = FALSE
)
}
\arguments{
\item{Y}{The index matrix}

\item{X}{The covariate matrix}

\item{B}{The initial coefficient matrix}

\item{lambda}{The regualrize coefficient}

\item{rank}{The rank of B matrix}

\item{alpha}{The step in the optimization process}

\item{control1}{A list that contains parameters for minimize loss function}

\item{control2}{A list that contains parameters for regularize}

\item{select.method}{link funtion selection method}

\item{plot}{whether or not plot the error during the iteration}

\item{descend.method}{optimization method}
}
\value{
A list that contains information such as the final B matrix, the training error.
}
\description{
The main function of mSIM package
}
\examples{
X <- scale(X_train)
Y <- scale(Y_train)
test = get_B_ADMM(Y=Y, X=X, lambda=0.2, rank=3, alpha=1, control1 = list(max.iter=5e1, tol=1e-5),
control2=list(ele.sparse=FALSE, row.sparse=TRUE, low.rank=TRUE), select.method='linear', plot=FALSE)
}
