% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict}
\alias{predict}
\title{The predict function for the mSIM package}
\usage{
predict(Y, X, B, Y.true, X.pred)
}
\arguments{
\item{Y}{The training index matrix}

\item{X}{The training covariate matrix}

\item{B}{The coefficient matrix}

\item{Y.true}{The groundtruth for predicting index matrix}

\item{X.pred}{The covariate matrix for prediction}
}
\value{
A list that contains the prediction result of Y and the MSE loss
}
\description{
The predict function for the mSIM package
}
\examples{
pred = predict(Y = Y, X = X, real.MSIM$B.sparse, Y.true = Y.true, X.pred = X.pred)
}
