% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X_test-Data.R
\name{X_test}
\alias{X_test}
\title{This is a dataset with 18 observations and each observation has 39 dimensions for testing.}
\description{
This is a dataset with 18 observations and each observation has 39 dimensions for testing.
}
