% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X_train-Data.R
\name{X_train}
\alias{X_train}
\title{This is a dataset with 100 observations and each observation has 39 dimensions.}
\description{
This is a dataset with 100 observations and each observation has 39 dimensions.
}
